+++ lib/local/LandmarkDetector/include/LandmarkDetectorModel.h
@@ -123,8 +123,8 @@ public:
 	cv::Mat_<double>		detected_landmarks;
 	
 	// The landmark detection likelihoods (combined and per patch expert)
-	double				model_likelihood;
-	cv::Mat_<double>		landmark_likelihoods;
+	float					model_likelihood;
+	cv::Mat_<float>		landmark_likelihoods;
 	
 	// Keeping track of how many frames the tracker has failed in so far when tracking in videos
 	// This is useful for knowing when to initialise and reinitialise tracking
@@ -194,8 +194,8 @@ private:
 	void NonVectorisedMeanShift_precalc_kde(cv::Mat_<float>& out_mean_shifts, const vector<cv::Mat_<float> >& patch_expert_responses, const cv::Mat_<float> &dxs, const cv::Mat_<float> &dys, int resp_size, float a, int scale, int view_id, map<int, cv::Mat_<float> >& mean_shifts);
 
 	// The actual model optimisation (update step), returns the model likelihood
-    double NU_RLMS(cv::Vec6d& final_global, cv::Mat_<double>& final_local, const vector<cv::Mat_<float> >& patch_expert_responses, const cv::Vec6d& initial_global, const cv::Mat_<double>& initial_local,
-		          const cv::Mat_<double>& base_shape, const cv::Matx22d& sim_img_to_ref, const cv::Matx22f& sim_ref_to_img, int resp_size, int view_idx, bool rigid, int scale, cv::Mat_<double>& landmark_lhoods, const FaceModelParameters& parameters, bool compute_lhood);
+    float NU_RLMS(cv::Vec6d& final_global, cv::Mat_<double>& final_local, const vector<cv::Mat_<float> >& patch_expert_responses, const cv::Vec6d& initial_global, const cv::Mat_<double>& initial_local,
+		          const cv::Mat_<double>& base_shape, const cv::Matx22d& sim_img_to_ref, const cv::Matx22f& sim_ref_to_img, int resp_size, int view_idx, bool rigid, int scale, cv::Mat_<float>& landmark_lhoods, const FaceModelParameters& parameters, bool compute_lhood);
 
 	// Generating the weight matrix for the Weighted least squares
 	void GetWeightMatrix(cv::Mat_<float>& WeightMatrix, int scale, int view_id, const FaceModelParameters& parameters);
